Get rid of Course vs CourseTemplate models and routes (#484)
* Rename database tables from base_course to course

* Blind rename of BaseCourse, base_course, and bcccv for models.

* Blind rename of BaseCourse, base_course, and bcccv for controllers

* Blind rename of BaseCourse, base_course, and bcccv for services

* Blind rename of BaseCourse, base_course, and bcccv for specs

* Blind rename of BaseCourse, base_course, and bcccv for javascripts

* Blind rename of BaseCourse, base_course, and bcccv for lib

* Blind rename of BaseCourse, base_course, and bcccv for views

* Rename BaseCoursesHelper to CoursesHelper

* Fix BaseCourse related routes.

* Fix BaseCourse related policies

* Fix BaseCourse in README

* Fix BaseCourse in db/dummies

* Make Course Managment page load.

* Update TODO with new input after discussion.

* Replace BaseCourse.type with BaseCourse.is_launched.

Also get the top-level Course Managment page to load.

Note: we're using an is_launched boolean field rather than
repurposing the `type` field into something that we tie logic
to. It's cleaner to tie specific functionality to a specific
boolean rather than having type drive the functionality and needing
to reverse engineer what types have what functionality.
Add more booleans for future features we want to tie to a particular
state of a course instead of over-engineering out of the gate.

* Get rid of "type" param in courses_controller

* Fix showing New Program Launch page

* Allow Waivers to be shown for a non-launched Course

* Fix CourseTemplate in db/dummies.rb

* Missed a file name for policies.

* Migrate CourseCustomContentVersion.type's

* Update note on CourseTemplates being editable from Course Mgmt show.

Just updating the language from Course to Launched Course since we
still haven't decided what the end user experience will be like for
non-launched courses. For now, they are all still treated like a
template.

* Get rid of template's in New Program Launch services

* Get rid of "type" in Course spec factory

* Get rid of course_template associations.

* Rename course_template to source_course for New Program Launch

Note: I left the language for the end user a "template" b/c I still
don't know what else to call them.

* Remove BaseCourse routing specs.

Not necessary to test routing here now that it's not
polymorphic based on type.

* Get rid of the rest of the course template spec factories.

Note: this is meant to completely break the specs. It's too much to
go through and fix them all in one commit, so that's up after.

* Get rid of polymorphic course paths

* Fix launch_program_spec

Mostly this was about an incorrect "presence" validation
for a boolean. When false, validation fails. But this also cleans
up the variables and factories to use things that make sense.

* Fix fetch_canvas_assignments_info_spec

* Fix course_spec.rb

* Fix sync_portal_enrollment_for_account_spec.rb

Don't know how these were ever working. Must have been a fluke
of magic matching IDs from elsewhere...

* Fix waivers_controller_spec.rb

* Fix waiver_submissions_controller_spec.rb

* Fix course_resources_controller_spec

* Make canvas_id default on the course factory.

It really should be a required field, but the scope of looking into
and making that change is too much. Just start by making all the specs
run on a course with a canvas_id.

* Fix courses_controller_spec.rb

Note: we only treat edit and destroy b/n launched and
unlaunched courses at the moment. Pull everything else out
of those contexts. We need to revisit what actions we allow
and what the behavior is once we're closer to going live
in order to prevent something accidental from happening.

Task: https://app.asana.com/0/1174274412967132/1199503585782989

* Fix peer_reviews_controller_spec.rb

* Fix course_custom_content_versions_controller_spec.rb

* Fix survey_submissions_controller_spec.rb

* Fix linked_in_authorize_spec.rb

I'm not sure what specifically about this PR caused this spec to break
b/c I think this was introduced with the CSS refactor:
https://github.com/bebraven/platform/pull/474 where I added styling
to the root application layout.

Anyway, I see this "You need to sign in or sign up before continuing."
alert all the time in dev when I'm refreshing or returning to a page
that is logged in, but something / somewhere timed out causing the
CAS stuff to go revalidate. This *should* fix that.

* Cleanup a couple remaining "template" references.

* Hide Edit and Delete for Launched Courses for the time being.

We need to think through the ramifications of allowing this
on a live course, so wait to bring the buttons back until we do that
thinking and prioritize projects to allow what we decide:
https://app.asana.com/0/1174274412967132/1199083857567510

* Get rid of course_management path

* Rename course_templates to unlaunched_courses.

The only place we use "template" now is in teh UI for designers.
The code is all launched vs unlaunched.