Make the Edit button easier to act on in Course Mgmt (#490)
* Make the Edit button easier to act on in Course Mgmt

The main thing we want to do it edit a course from the
Course Managment page. I kept accidentally hitting the name of
the course every time which brought me to it. This changes the layout
to make Edit way more prominent and closer to where your eye
is visually when looking at the course to edit.

It's also a step in the direction of showing a list of Programs in
the second section instead of Launched Courses. The top section
with Templates will be one way to launch a new program, by choose the
templates. Eventually, the second way will be to Re-Launch an existing
program from the list in the second section or on the individual program
page that each item brings you to showing the courses in that program.

* Remove the badge since they're all a Course now.