Get rid of "Sign-in or sign-up" flash alert. (#492)
This is pretty hacky but I could spend days going through and
trying to understand and fixup how the alert, devise, and CAS
all interact. Suffice it to say that b/c of SSO where we redirect
to the sign-in page and b/c there is no "sign-up" option, this message
is safe to hide and is annoying.