Add some CSS styling for Peer Reviews. (#489)
* Add some CSS styling for Peer Reviews.

This is just a first pass, but I'm pretty happy with it. We used the
pinned screenshot in this task: https://app.asana.com/0/1174274412967132/1193980665758237
as inspiration. This was from a prototype we did in FormAssembly.

One thing to note is that this CSS is very specific to this form.
It was a challenge to get the radio list with 5 options and inline
text to lay out nice horizontally and switch to vertical at the
right media breakpoints. See the big comment in
`peer_review_submissions.scss` for why we have a weird hard-coded
non-standard media breakpoint in there. I did pretty extensively check
how the form looked at different screen sizes when shown inside Canvas.

* Address code review feedback.

- Get rid of headings in the legends
- Use fieldset/legend instead of aria role stuff for the individual questions
  to group the radio options
- Now that headings are gone, we don't mess with the capitalization of the text
-