Ability to manage Canvas enrollments from User's Admin dashboard (#499)
* Prototype of adding "Enrollments" to the Users Admin page

This bypasses Salesforce and lets you setup a user in a
course in a section as a certain role from teh Admin Users
page.

* Cutover UserSections to UsersRoles.

We introduced the user_sections table before we go into actually
implementing stuff and never used it. What we actually have for this
is the polymorphic users_roles table that the Rollify gem gives us.
That's logically what this controller is meant for. It's to edit
the users_roles.

* Show list of existing Enrollments.

* Make more room for the Enrollments table.

* Add route to delete enrollment

* Implement deleting an Enrollment

* Remove enrollments editing from the New User page (for now).

We'll likely want to be able to provision a new user in Canvas and
set their enrollment info. But for now, make this only work on
existing Users since it's broken for new ones.

* Make the Enrollments tool only work for unlaunched courses.

See the big note in app/views/users_roles/_new.html.erb for why.

* Fix TaEnrollments on Users Admin page.

The Sync To LMS logic enrolls a TA in both the Fellow Accelerator course
and the LC Playbook course. Needed to mimic that in the enrollments
UI.

* Add "No enrollments" row to list if there are none.

* Change placement of Edit Enrollments button next to heading.

* Don't show Canvas section for adding new users.

Adds a big note about why we don't support adding a new
user to Canvas as well using the Users Admin page.

* Add a big comment about what users_roles_controller is

* Add a policy for UsersRolesController

* Make code clearer

* Remove unused index route

* Add basic specs.

This is Admin only and written a bit hackily. We're likely to keep
iterating and it's not worth writing and rewriting specs for hacky
code. I just wanted something basic to detect if the
`sync_portal_enrollment_for_account` stuff gets refactored and breaks
this so that the dev is aware that this feature exists.

* Cleanup comment

* Code review feedback.