Remove LRS, use ProjectSubmissionAnswer (#482)
* Hax. Remove LRS, use ProjectSubmissionAnswer

* A bunch of stuff

* Commit work so i can switch to the other branch

* Fix migration, answers query

* Implement answer prefill

* Refactor

* Fix stuff

* Fix logic

* Fix answers not saving.

We still have an issue with input ids, but this at least gets
past all the network call stuff and makes sure we have a valid
project submission we're tying answers to.

* Back out changes to project_submission#new.

We had decided to create the new submission when the first answer came in.
I conflated issues when making the previous change. It was failing for a
different reason and also complaining about something that wasn't really a problem.
This goes back to the original implementation of project_submission#new
and gets rid of the console log about no project submission when trying to
pre-fill. We don't need to prefil if not answers have already been saved to a submission

* Refactor and add first bit of Honeycomb tracing to xapi_assignment.js

* Add Honeycomb tracing for saving answers

* whitespace/typo

* Fix policies?

* Proposal for alternate way to do ProjectSubmission/Answers routing

* Clarify comments.

* Fix answer index/prefill

* Remove allow_multiple_submissions

* Submit the correct thing (manual grading required).

I spent way too long trying to handle error cases and keep things
consistent. Leaving in some hints for future devs to save them the
headache of even trying.

The only real downside effect of a failure with this code is that Canvas
will show "Re-Submit" even though there is no submission visible there.
To solve that, my best idea is to start saving the canvas_submission_id
on the project_submission. Update that after we successfully create it
and use that in addition to `is_submitted` to determine if the button should
show Submit or Re-Submit.

* Fix highlighting the answers on submitted submissions.

* Rename xapi_assignment to project_answers and fix bug.

The bug was that I was highlighting user input even for unsubmitted
submissions.

* Cleanup the Honeycomb field names to display traces nicely.

* Don't save answers for already submitted ProjectSubmissions

* Update submission model

* comment

* Fix syntax

* Remove validation

* Remove highlighted-user-input class

* Remove unused submissions columns

* Add factories

* Fix project_submission model specs

* Add answer model specs

* Fix project_submissions_controller_spec#show

* Remove login specs from project_submissions_controller_spec

We should just test this stuff as part of application_controller specs,
not in every individual controller. Too hard to maintain.

* Refactor project_submissions_controller_spec to be more DRY

* Fix project_submissions_controller_spec#submit

* Add rails-controller-testing gem for `assigns` utility.

See the comment in 'first time viewing submission' for why
we need this.

* Implement project_submissions_controller_spec#new

* Implement project_submissions_controller_spec#edit

* Fix policy specs

* Add answers controller spec

* Start jest specs

See comment for article on how to do it

* Comment out project_answers specs (formally xapi_assignment).

Need to reimplement these but we also need to get this branch deployed
to prod for some UX testing sessions today. Implement these in a follow-up
PR.

* Add 'jest-fetch-mock' package in preparation for Jest tests.

Haven't actually implemented anything real yet, but we need one
test in there that doesn't fail in order to merge this and `fetch`
needs to be defined for the project_answers.js to run. We're planning
on using jest-fetch-mock to test this, so this just gets that in place.

* For consistency, add uniqueness contraints to ProjectSubmission and ProjectSubmissionAnswers.

To do this, we have to stop using create_or_find_by!() and instead rely on the
database constraints. I think it's cleaner this way?

The answers were failing to save b/c the uniqueness constraint wasn't
playing nicely.

Co-authored-by: rshipp <github@rshipp.com>
Co-authored-by: Brian Sadler <sadler.brian@gmail.com>