Add a New Course Template initialization process (#509)
* Update UI for how a new Course Template will be initialized.

This changes the primary links for both course and Projects/Surveys
to show platform view of it instead of bringing you to Canvas. When
using the Course Management stuff, the primary goals is platform related
management and not viewing the content on Canvas. A new "View On Canvas"
link as a secondary action has been added.

* Bug in New Program Launch after Course/Course Template refactor.

* Make "View On Canvas" links open in a new tab.

* Implement "Initialize New Course Template job.

TODO: refactor Launch New Program to use this new CloneCourse service.

* Fix courses#new spec

* Fix courses_controller#create specs.

* Write services/clone_course specs.

* Refactor LaunchProgram to use CloneCourse service.

Note: most of the logic of the specs had already been cutover
to the clone_course specs in the previous commit.

* Add specs for clone_course_job

* Small cleanup of comments and new_params.

* Fix language on courses#show page.

* Cleanup links to Canvas.

The Rise360Modules used the old layout of the name linking to Canvas.
Switch that to use the new layout of the name linking to the edit page.
Also, get rid of the Canvas links opening in a new tab b/c it's bad
practive and people can just Cmd-Click if that's their preferred workflow.

* Code review feedback - un-nest `create_from_course_id` param.

When using the `form.hidden_field` Rails sticks it inside the model
the form is for. It's a bit confusing since it's not a column on that
model so just hand-write the input field (we do this elsewhere for passing
the STI type in).