Consolidate AccountCreator and SetupPortalAccountJob into RegisterUserAccount (#513)
* Remove obsolete SyncToLMS code.

As part of the following PR, the old lib/sync_to_lms.rb class
was deprecated. This removes all the old code (and switches the
one last obsolete reference over to the new code).

https://github.com/bebraven/platform/pull/337

* Typo

* Consolidate AccountCreator and SetupPortalAccountJob into RegisterUserAccount

When a user registers using a link like:
https://platformweb/users/sign_up?u=0031J00001y1vQG

it's responsible for creating a new Canvas account, enrolling the
user in the proper Canvas section, updating Salesforce to match,
and creating the local User record.

This logic was very hard to follow and there was a lot of obsolete stuff
in there for Highlander like DocuSign and syncing to the Join server.
This cleans up this flow to make it easier to understand in preparation
for changing the Sync part of this to leave them in the CohortSchedule
section when running the logic to adjust their enrollment when changes happen.

TODO: fix specs and add rollback code if the RegisterUserAccount logic
fails partially.

* Remove Sync To Join code.

This was a hack to temporarily continue using the old Braven Network.
The plan is to cut it over and get rid of Join before Highlander launches.

* Remove DocuSign code. It's been replaced with FormAssembly

* Remove the rest of the Sync To Join code.

* Remove obsolete AccountCreator spec.

* Fix register_user_account_spec

* Remove TODO about the Salesforce field for canvas user ids.

I verified that there isn't a unique constraint on teh SF side
so it won't cause failures. I also remember having a convo with
Robert about this and we decided it was fine to share the field
but I can't find the convo/task to link.

If there are issues here, let's tackle them when we find them
instead of littering the code with TODO's that may not actually
be something to do.

* Missed another Sync Join file

* Error handling

* Reverse creation order, remove fallback code

* Add more tests

* Change comment

Co-authored-by: rshipp <github@rshipp.com>