Only send one confirmation email. (#558)
When I refactored the services into RegisterUserAccount, I introduced
a bug where the confirmation email was being sent twice. Once when the
new User model was created and once by explicitly calling the method to
send. This just relies on the User model creation to have it go out.

Note: accidentally using the sign-up form twice is untested. I'll
queue up a future task to sanity test that.

Test Plan:
- bundle exec rspec spec/services/register_user_account_spec.rb