Make sure and grade the proper Rise360ModuleVersion (#560)
* Use most recent Rise360ModuleVersion for grading

* Handle Module grading when there are no mastery questions

* Add language about messing up grades if you Publish Latest

* Use canvas_assignment_id to find the Rise360ModuleVersion to grade

If we just grab the last Rise360ModuleVersion using it's activity_id,
we may grab the one from the incorrect course which could have been
published later with a different number of quiz questions. We need to
make sure and get the proper module version corresponding to the
Rise360ModuleInteraction that we're grading. The canvas_assignment_id
let's us do that since we always create or update the
CourseRise360ModuleVersion which joins the course, assignment, and module
version together when we publish or publish latest.

* Add some logging in the grade:modules task

Rake tasks don't seem to be able to use the Rails logger,
but if we "puts" to STDOUT, you can see it both on the command
line and in the Heroku logs (which use STDOUT). This makes it
easier to troubleshoot.

* Add unique constraint for canvas_assignment_id in CourseRise360ModuleVersion.

The join table from the course to the module version logically represents
a Canvas assignment. There can only be one.

* Missed a spot with canvas_assignment_id in the specs

Co-authored-by: Brian Sadler <sadler.brian@gmail.com>