Fix errors seen by Canvas users (non-admin) when ending up on Platform. (#575)
* Fix errors seen by Canvas users (non-admin) when ending up on Platform.

Test Plan:
- Change your password, the Log In button should bring you to /cas/login endpoint
- Try to sign-up an account that is already signed up, the Log In link should be /cas/login
- Logout of Canvas (in both Chrome Incognito and normal), the Log In link should be /cas/login
- At the /cas/login page, logging in as a non-admin should bring you to Canvas regardless
  of the service parameter (which will be to platform)
- Logging in as an admin should bring you to Platform

Note: in Chrome incognito, Canvas was sending a `url` parameter as the `continue_url`
which lead to a non-existant `logout` view. Logout should always render the `login`
page.

* Fix custom_contents/smoke_spec.rb

This is a staff page so we don't care that logging in directly to the URL
brings you to the root of the site instead of back to the Content Editor.
Just explicitly visit the page we are testing after logging in.

* Fix linked_in_authorize_spec.

This is always launched from within Canvas, so switch to using the
LtiLaunch state for authorization.