Fix Waivers and Pre-Accelerator Survey "Completed" CSS (#586)
The "Completed" template was written to look nice in the lti_canvas
layout. We introduced a form_assembly layout and changed pre/post-accelerator
surveys flow to be outside of Canvas. In doing both of these, we missed
a couple spots where we should use lti_canvas as the layout instead
of the form_assembly one.

Note that it's a bit hacket that it uses the lti_canvas layout when not
shown inside Canvas for the `create` actions, but it didn't seem worth
trying to make cleaner.