Inject CSS into the Rise360 content (#588)
* Put Rise360 Module in iFrame and inject our own stylesheet into it.

CSS in a parent frame doesn't apply to a child iframe, but if you're
on the same domain you can use Javascript to get inside the iframe
and do stuff. This commit moves the Rise360 content inside an iframe
instead of just proxying the package directly from S3 and then injects
our own CSS stylesheet into it so that we can apply custom styles to
Rise360 content.

* Refactor to remove rise360_module layout.

I was hacking on stuff when i added that, but we really don't
need it. It's cleaner this way to have the CSS that we inject
into the Rise360 content have ONLY CSS for that content. Now
the CSS for the container and iframe itself is done in a more
standard Rails way by just being in the lti_canvas layout's CSS file.

* Fix specs and change path from absolute to relative.

Since this will only work on the same domain, seems safer to
use the relative path when iframing in the Rise360 content

* Accidental commit.

* Left some troubleshooting code in