Make "Rate This Module" look good on mobile. (#603)
* Rough pass of making Rate This Module look good on mobile.

There are a bunch of TODOs left in here, but this does a pretty
good job on both the smallest screen size and biggest screen size
of making the RTM view look just like it's normal Rise360 content.

* Add some media queries to style Rate This Module at various screen sizes.

Note: this is just a "good enough" pass at this and it still looks
not great on a bunch of screen sizes. See the novel length comment in
rate_this_module.scss. The good news here is that the odds of a scrollbar
showing up is pretty low. The bad news is that the content doesn't always
line up with the rest of the Rise360 content.

* Add the ability to force a Honeycomb Boomerang beacon to go out.

This Rate This Module stuff runs after the page is loaded so
the fields added don't make it in the page load beacon. Need to
explicitly tell Boomerang to send out a beacon measuring this
event.

* Cleanup the Javascript that adds rate-this-module CSS classes to Rise360.

* Remove aria-label in preference for actual label.

Styling the question prompts so they were distinguished from the
overall intro paragraph was challenging using paragraph text. Switched
to labels which would conflict with the aria-label.

* Whitespace

* Add example HTML in lieu of Jest tests.

I tried using jest-erb-transformer to get the .erb Javascript processed
and transformed so we could use it with Jest but it wasn't worth the time.
Just put the sample HTML in there so that if it breaks one day (which we
don't control so Jest was just serving as a specification for how it currently looks)
then this is a convenient reference.

* Fix Jest tests for now Honeycomb method.

* Link Asana task for TODO