Show something to Fellows when they view their Module grade (#607)
* Add method to lti_score for Modules to use to create a new score/submission

* When folks start working on a Module, get a placeholder submission in place.

We're going to use the submission to launch our LTI endpoint that shows
them helpful information about their Module grade in order to reduce
questions to staff and support requests.

NOTE: when we deploy this, all previously graded Modules may have their
grades zeroed out when folks open the Module again.

* Rough pass of getting this working.

After discussion, going to refactor the controller and routes
used a bit

* Refactor from course_rise360_module_versions_controller#show

Now the view is on rise360_module_interactions_controller#index
which is nested under a CourseRise360ModuleVersion

* Fix headings for screenreaders and clarify calculation formula.

Some students really want a detailed breakdown of how their grade is
calculated. So get super clear when breaking down the calculation to
help them understand without asking for staff support.

* Add an authorization policy

* Handle TA's and LC's viewing Modules.

It would fail before if you weren't enrolled as a Student. Now
TAs and LCs can view a module, it just won't create a submission.

Note: currently I think the grading code doesn't filter out non-students.
Need to do that: https://app.asana.com/0/1174274412967132/1199946751486950

* Make it look better on mobile

* Show message when the grade is not up to date.

Note: I ran into a major hiccup here b/c I was assuming we could
get the interactions from information we had in the lti_launch,
but that's not the case when a TA or LC is trying to view a Fellow's
submissions. Need to refactor.

* Rename CSS classes to change the font size to be less confusing.

* Add a Rise360ModuleGrade model.

* Refactor to use Rise360ModuleGrade.

This is so that TA's and LC's can view the grade explanation page of
a fellow. We need to store who the grade is for and not just assume it's
the grade of the user viewing it.

* Add user/canvas_assignment_id index to rise360_module_interactions.

This is how we query when determining if their grade is up to date.

* Write the policy

* Remove old code from before refactor

* Some cleanup of comments

* Specs for Rise360ModuleVersionsController

* Add specs for Rise360ModuleGradesController

* Add Rise360ModuleGradePolicy spec

* Code review feedback