Pin postgres to v12 (#624)
There don't seem to be any major breaking changes in v12.0-v12.6 (the
current latest v12 as of this commit), so I'm assuming it's relatively
safe to pin to the major version. If we run into problems migrating
in the future, we may want to reconsider and pin to a specific minor
versionm, at the cost of having to periodically check and upgrade
minor versions manually.