Filter sensitive data out of our logs (inc. Honeycomb and Sentry) (#618)
* Filter out state param from various params.

Note that I think I'm going to have to refactor this to use
https://api.rubyonrails.org/classes/ActionDispatch/Http/FilterParameters.html
instead b/c the below doesn't handle filtering out the query strings in
logs like: Started GET <blah_path>/state=the_value

* Move parameter filtering back to an initialize + add redirect filtering

This is the default location to do this for a Rails 6 app, so leave it there.
Plus, I'm working on using filter_attributes to deal with the database
SQL query logging and want to put this here as well. It let's us consilidate
all filtering logic to one file.

* Rough implementation of filtering params, Honeycomb data, and database queries.

I need to clean this up and do a lot more testing to make sure I got everything

* Refactor filtering logic into helper class.

This will make it easier to test.

* Filter "Redirected to ..." logs properly.

This makes the filter only apply to the sensitive params instead
of filtering the whole path.

* Filter sensitive data from ActiveRecord logs of the SQL statements

* Remove some cluttery comments. Will move the examples to specs

* Add TODO in restclient_instrumentation

* Remove troubleshooting code

* Add Sentry TODO so I don't forget.

* SQL attr name could be nil. Handle that.

* Add some basic specs

* Add more database columns to filter. Share code with params_filter too.

* Make really messy code slightly less messy

* Upgrade Sentry gems.

More info here: https://docs.sentry.io/platforms/ruby/migration/

* Move the FilterLogging.is_enabled? check inside the class to make it cleaner.

* Proof of Concept filtering Sentry data.

This works for the session cookie, the referer header, and
and the query string. It also shows how you *could* filter the
breadcrumbs, but it's not a very good solution. The next iteration
will be to try and filter that at the root, when ActiveRecord sends
the event notification

* Refactor SQL log filtering to be before the ActiveRecord::Notification goes out.

This way both Senty and Honeycomb only see the filtered value since they construct
their payloads by subscribing to the notifications.

* Fix some Honeycomb filtering I messed up trying to cleanup the code.

Apparently, the inline filter wasn't working when I thought it was.
Need to actually store the filtered value back in the hash.

* Filter Sentry breadcrumbs + better error handling.

* Move RestClient filtering to the consolidated FilterLogging class.

Also, rename the fields to match what we discussed here:
https://app.asana.com/0/1174274412967132/1200055379602516

* Make the code more obvious what is happening

* Fix specs.

* More specs. Cleaning up TODOs.

* Cleanup moar TODOs.

* Remove unecessary regex escape.

* Fix RestClientInstrumentation specs

* Fix HoneycombJsController specs

* Disable Sentry for tests

* Filter sensitive data in all environments

This way it's more likely we catch something slipping through
and we should be able to get whatever we need in the dev env
from the local database or browser console.

* Don't exclude a few more Sentry exceptions

If we see these bombard Sentry in the wild, we can start excluding
them from being sent but out of the gate lets send as much as possible
to Sentry

* Filter id_token_payload column too.

This is an encrypted payload (signed JWT) containing information about the user,
launch, context, etc. It would take some doing, but a nefarious actor could decrypt it
using Canvas public JWKs

* Missed a bunch of CAS login related tickets when filtering.