Remove obsolete login code (Join server CAS, etc) (#641)
* Get rid of obsolete login code.

This originally came from the Join server and the login form would
also be embedded inside an iframe sometimes where we didn't want the
header and footer to show. We also had some custom stuff in our old
Portal where we added some code to tell the Join server to log itself
out instead of using the built-in CAS "sign-out everywhere functionality"
which may or may not work. Either way, all that stuff is obsolete now
so this rips it out.

* Remove ability to configure server to auth again Join.

Platform CAS SSO is stable and we're getting rid of the Join server.
There is no chance we configure the Highlander platform to auth
against Join.

* Remove ENV['BZ_AUTH_SERVER'] from all specs.

We had to run specs differently for the old Join path
vs the new local platform SSO CAS path. Rip all that out
since everything will be local platform SSO going forward.

* Missed a spot

* Missed a file. This was the old CAS Authenticator.

* Remove infoline.

It's not currently showing on the page and their license doesn't
require it: https://github.com/rubycas/rubycas-server/blob/master/LICENSE

* One step closer to a standard Rails form.

I started trying to replace the fields and inputs with tag helpers,
but it wasn't worth the effort. The main thing here is that this
gets rid of the explicit authenticity token stuff and lets the
`form_tag` take care of that.

* Fix specs