Rate This Module now shows an indication that it was submitted. (#643)
* Rate This Module now shows an indication that it was submitted.

This is a bit hacky. Flash notices don't work if we don't have
session, which we often don't when shown inside an iframe inside
Canvas cross domain (esp in Safari or Chrome incognito). Instead,
just pass the info about the flash notice through in query params
and also switch the button text.

The reason I had to use two query params is that that the Close of
the flash actually redirects back to edit again and we want to keep
the Re-Submit text.

I didn't think it was worth dealing with showing the Re-Submit text
on subsequent launches of the Module. It will pre-load the old response,
but "Re-Submit" only shows when you're working on it inline.

* Fix specs

* Add submission indication status specs

* Upgrade mimemagic gem from 0.3.5 to 0.3.6

They release 0.3.6 today and yanked 0.3.5 off https://rubygems.org/
so the build is failing