Better UX for signup / registration / confirmation flow. (#642)
* Better UX for signup / registration flow.

1. If they signup but then try and use the link a second time,
   this change brings them to the login form where they just have
   to type their password.
2. If they try to login before confirmation their email address,
   this change brings them to the page telling them to do that
   with a button to resend the confirmation link if they can't find it.

Tasks:
- https://app.asana.com/0/1174274412967132/1199689504093092
- https://app.asana.com/0/1174274412967132/1199575667690414

* Clarify some comments

* Specs for new login page with only a password field.

Note that I did one more pass of cleaning up the HTML as part of this
b/c I was trying to test that the password field got autofocus so
I was tinkering with the attribute. The tabindex stuff was unecessary
b/c inputs can get focus by default and their order is how they show
up in the source. Maybe this was needed before when there were focus
elements in the header that we didn't want to be first?

* Cleanup prefix specs ever so slightly

* Add specs for unconfirmed user login.

* Login should be "log in"