More Honeycomb field normalization (#649)
* Add prefixing the field names with the "controller" back in.

This makes it so that you don't have to manually prefix all the field names,
just construct the HoneycombSpan and all addField() calls will prefix them.

Also, make the name, controller, and action come through. When unprefixed,
some rails magic was masking them with the Rails auto-instrumented values,
like honeycomb_js and send_span. By adding the js.app prefix, they come through.

* Rename "controller" and "action" to "fileName" and "functionName"

Folks just getting started on the codebase will have an easier time
using this wrapper correctly if the params are named according to what
should go in them instead of having to think about the bigger picture
concept of Rails controllers and action vs JS stuff.

* Add the ability to turn off fileName prefixing.

This should be relatively rare and only used for common, generic
fields that are not specific to the specific JS being instrumented.
Most of these types of things will be set in the Rails code though
and not need to be passed through from JS.

* Fix field naming for project_answers.

1. No need to send the api_url, it never changes.
2. Use snake_case for fields that don't represent database columns.
3. Don't prefix the response.status field with the fileName. This is a generic field not specific to this controller.

* Switch project_answers from snake_case to CamelCase for JS code.

* Switch rise360_module_versions from snake_case to CamelCase for JS code.

* Get rid of some params that shouldn't be sent as part of the JS span.

See the comments in PropagatedSpan.new() where we loop over the params
for more details.

* Move the information about the type of beacon to its own field.

The name of the span should always be what the dev chooses. If it's
an auto-instrumented boomerang beacon/span, just call it js.boomerang.
Move whether its a page load/unload or an XHR beacon into a js.boomerang.beacon_type
field.

* Accidentally CamelCased some things that were actually JSON.

They needed to be snake_case b/c that's what the Rails controller works with

* Move the "page_load" and "page_unload" beacon info to a field.

Always just send the name that the JS side sends b/c it should
tell where the info is coming from if we're using the HoneycombSpan
wrappers. If the beacon didn't have any manual instrumentation
(aka no name set), just set it to js.boomerang.

* Add app.user.id to Honeycomb for every controller span.

The intent here is to make it easier to correlate exceptions
to users, but I'm finding it hard to test exception in dev. They
don't see to set the "error" and "error_detail" fields in the root
span like the do in prod. Actually, I couldn't see them set in any spans,
BUT this does put the 'app.user.id' in the same span that has a
'process_action.action_controller.exception' so even if it doesn't end
up in error / error_detail, that should make querying for exceptions
and correlating them to users easier.

* Cleanup 'name' field. It's always used now so we don't need js.app.name.

* Switch the span name back to being normalized. Use "context" instead.

There may be multiple HoneycombSpanBase's in a single actual span
sent to Honeycomb. This is why the page load and unload beacon names
where always overwritten and normalized. This cleans this all up so
that the span names are always a normalized name depending on what
type of beacon it is. The extra information about where in the code
the instrumentation we're adding is coming from is now sent as the
prefix for all fields by default and also set as a standalone field
set to true.

Rename HoneycombSpan to HoneycomabAddToSpan to make this more clear.
Let the HoneycombXhrSpan name alone b/c that is a 1-1 mapping from
JS to the actual span sent to Honeycomb since it's wrapping a Fetch/XHR
call.

* Fix specs. Add missing HoneycombAddToSpan error spec and sendNow() spec

* Remove the 'context' as a standalone field.

Originally, I through this could be used to query for all spans for a particular
HoneycombSpanBase and if generic, non-prefixed fields were added, you could tie
the two together. However, async stuff can happen after page load so those fields
will go out in the next beacon. Just always query for the fields themselves and
you'll get an auto-complete list of what's available in the timeframe you're
querying against.

* Manually sent beacons can also come through with "xhr.pg" param.

Before this, that was recorded as a page load, but with this change
when we loop for up to 10 seconds and then manually send a beacon,
it gets the right js.boomerang.manual name.

* Fix honeycomb_js_controller_spec

* Fix specs (hacky).

I tried everything I could think of and the spec failure are unrelated to this
change, it's just VCR weirdness with the network call which should be allowed
and we have a CI dataset that specs send data too. Punting on trying to get to
the root cause.

* Fix comment.